+++ src/main/java/me/efekos/simpler/commands/CoreCommand.java
@@ -29,7 +29,6 @@ import me.efekos.simpler.commands.syntax.ArgumentPriority;
 import me.efekos.simpler.config.MessageConfiguration;
 import me.efekos.simpler.exception.InvalidAnnotationException;
 import me.efekos.simpler.translation.TranslateManager;
-import org.bukkit.ChatColor;
 import org.bukkit.Location;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
@@ -39,13 +38,15 @@ import org.jetbrains.annotations.NotNull;
 
 import javax.annotation.Nullable;
 import java.lang.reflect.Constructor;
+import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 /**
- * Used for core commands like /friends invite,/friends list,/friends remove etc. {@link #getSubs()} will return a list of the {@link SubCommand}s that belong to this command. Must be annotated with {@link me.efekos.simpler.annotations.Command} to be registered properly.
+ * Used for core commands like /friends invite,/friends list,/friends remove etc. {@link #getSubs()} will return a list of the {@link SubCommand}s that belong to this command. Must be annotated with {@link me.efekos.simpler.commands.Command} to be registered properly.
  */
 public abstract class CoreCommand extends Command {
     protected CoreCommand(@NotNull String name) {
@@ -62,8 +63,8 @@ public abstract class CoreCommand extends Command {
     @Override
     @NotNull
     public String getName() {
-        me.efekos.simpler.annotations.Command command = this.getClass().getAnnotation(me.efekos.simpler.annotations.Command.class);
-        if(command!=null)return command.name();
+        me.efekos.simpler.commands.Command command = this.getClass().getAnnotation(me.efekos.simpler.commands.Command.class);
+        if (command != null) return command.name();
         return super.getName();
     }
 
@@ -73,8 +74,8 @@ public abstract class CoreCommand extends Command {
     @Override
     @Nullable
     public String getPermission() {
-        me.efekos.simpler.annotations.Command command = this.getClass().getAnnotation(me.efekos.simpler.annotations.Command.class);
-        if(command!=null)return command.permission();
+        me.efekos.simpler.commands.Command command = this.getClass().getAnnotation(me.efekos.simpler.commands.Command.class);
+        if (command != null) return command.permission();
         return super.getPermission();
     }
 
@@ -84,8 +85,8 @@ public abstract class CoreCommand extends Command {
     @Override
     @NotNull
     public String getDescription() {
-        me.efekos.simpler.annotations.Command command = this.getClass().getAnnotation(me.efekos.simpler.annotations.Command.class);
-        if(command!=null)return command.description();
+        me.efekos.simpler.commands.Command command = this.getClass().getAnnotation(me.efekos.simpler.commands.Command.class);
+        if (command != null) return command.description();
         return super.getDescription();
     }
 
@@ -97,11 +98,12 @@ public abstract class CoreCommand extends Command {
     @NotNull
     @Override
     public String getUsage() {
-        return "/"+getName()+" <sub> <args>";
+        return "/" + getName() + " <sub> <args>";
     }
 
     /**
      * Returns a list of the {@link SubCommand}s that belong to this {@link CoreCommand}.
+     *
      * @return List of the subs.
      */
     @NotNull
@@ -109,14 +111,15 @@ public abstract class CoreCommand extends Command {
 
     /**
      * Finds a {@link SubCommand} by its name using {@link SubCommand#getName()}.
+     *
      * @param name Name of the {@link SubCommand} you want to get.
      * @return {@link SubCommand} if found, null otherwise.
      */
     @Nullable
-    public Class<? extends SubCommand> getSub(String name){
-        for (Class<? extends SubCommand> sub:getSubs()){
-            me.efekos.simpler.annotations.Command command = sub.getAnnotation(me.efekos.simpler.annotations.Command.class);
-            if(command.name().equals(name)){
+    public Class<? extends SubCommand> getSub(String name) {
+        for (Class<? extends SubCommand> sub : getSubs()) {
+            me.efekos.simpler.commands.Command command = sub.getAnnotation(me.efekos.simpler.commands.Command.class);
+            if (command.name().equals(name)) {
                 return sub;
             }
         }
@@ -124,182 +127,167 @@ public abstract class CoreCommand extends Command {
     }
 
     /**
-     * Grabs the value of {@link me.efekos.simpler.annotations.Command#playerOnly()} and returns it.
+     * Grabs the value of {@link me.efekos.simpler.commands.Command#playerOnly()} and returns it.
+     *
      * @return Is this command or subs of this command can be used by something that is not player?
      */
-    public boolean isPlayerOnly(){
-        me.efekos.simpler.annotations.Command command = this.getClass().getAnnotation(me.efekos.simpler.annotations.Command.class);
-        if(command!=null)return command.playerOnly();
+    public boolean isPlayerOnly() {
+        me.efekos.simpler.commands.Command command = this.getClass().getAnnotation(me.efekos.simpler.commands.Command.class);
+        if (command != null) return command.playerOnly();
         else return false;
     }
 
     /**
      * Provides a text that will help the {@link CommandSender} trying to execute one of the {@link SubCommand}s at this command.
-     * @param sender The sender who is struggling with using any sub of this command.
+     *
+     * @param sender       The sender who is struggling with using any sub of this command.
      * @param subInstances Instances of all the {@link SubCommand}s under this {@link CoreCommand}.
      */
-    public abstract void renderHelpList(CommandSender sender,List<SubCommand> subInstances);
+    public abstract void renderHelpList(CommandSender sender, List<SubCommand> subInstances);
 
     @Override
     public boolean execute(@NotNull CommandSender sender, @NotNull String commandLabel, @NotNull String[] args) {
         MessageConfiguration configuration = Simpler.getMessageConfiguration();
 
-        if(args.length==0){
-            sender.sendMessage(TranslateManager.translateColors(configuration.USAGE.replace("%usage%",getUsage())));
+        if (args.length == 0) {
+            sender.sendMessage(TranslateManager.translateColors(configuration.USAGE.replace("%usage%", getUsage())));
             return true;
         }
+
         Class<? extends SubCommand> cmd = getSub(args[0]);
-        if(cmd ==null || args[0].equals("help")){
+        if (cmd == null || args[0].equals("help")) {
             List<SubCommand> subCommands = new ArrayList<>();
-            for(Class<? extends SubCommand> sub:getSubs()){
+            for (Class<? extends SubCommand> sub : getSubs()) {
                 try {
                     Constructor<? extends SubCommand> constructor = sub.getConstructor(String.class);
-                    me.efekos.simpler.annotations.Command commandA = sub.getAnnotation(me.efekos.simpler.annotations.Command.class);
+                    me.efekos.simpler.commands.Command commandA = sub.getAnnotation(me.efekos.simpler.commands.Command.class);
                     SubCommand command = constructor.newInstance(commandA.name());
                     subCommands.add(command);
-                } catch (Exception e){
-                    e.printStackTrace();
-                }
+                } catch (Exception ignored) {}
             }
 
-            renderHelpList(sender,subCommands);
+            renderHelpList(sender, subCommands);
             return true;
         }
-        me.efekos.simpler.annotations.Command cmdA = cmd.getAnnotation(me.efekos.simpler.annotations.Command.class);
-        if(cmdA==null) {
+
+        me.efekos.simpler.commands.Command cmdA = cmd.getAnnotation(me.efekos.simpler.commands.Command.class);
+        if (cmdA == null) {
             try {
                 throw new InvalidAnnotationException(cmd.getName() + " Must have a @Command annotation.");
-            } catch (Exception e){
+            } catch (Exception e) {
                 e.printStackTrace();
             }
         }
-        String[] subArgs = Arrays.copyOfRange(args,1,args.length);
+        String[] subArgs = Arrays.copyOfRange(args, 1, args.length);
 
-        try{
-            if(sender instanceof Player){ //sender is a player
-                me.efekos.simpler.annotations.Command command = this.getClass().getAnnotation(me.efekos.simpler.annotations.Command.class);
-                Player p = (Player) sender;
+        try {
+            if (sender instanceof Player) { //sender is a player
+                me.efekos.simpler.commands.Command command = this.getClass().getAnnotation(me.efekos.simpler.commands.Command.class);
 
-                if(!command.permission().isEmpty() &&!p.hasPermission(command.permission())){ // @Command has a permission and player don't have the permission
-
-                    p.sendMessage(TranslateManager.translateColors(configuration.NO_PERMISSION));
+                if (hasPermission() && !sender.hasPermission(command.permission())) { // @Command has a permission and player don't have the permission
+                    sender.sendMessage(TranslateManager.translateColors(configuration.NO_PERMISSION));
                     return true;
                 }
+
                 // @Command don't have a permission or player has the permission
-                    if(getSub(args[0])!=null){
-                        if(!cmdA.permission().isEmpty() &&!p.hasPermission(cmdA.permission())){ // SubCommand's @Command has a permisison and player don't have the permisson
-                            p.sendMessage(TranslateManager.translateColors(configuration.NO_PERMISSION));
-                        } else { // SubCommand's @Command don't have a permission or player has the permisson
+                if (getSub(args[0]) == null) return true;
 
-                            a:{
-                                Constructor<? extends SubCommand> constructor = cmd.getConstructor(String.class);
-                                constructor.setAccessible(true);
-                                SubCommand instance = constructor.newInstance(cmdA.name());
-
-                                if(subArgs.length==0&&instance.getSyntax().getArguments().stream().anyMatch(argument -> argument.getPriority()==ArgumentPriority.REQUIRED)){
-                                    sender.sendMessage(TranslateManager.translateColors(configuration.USAGE.replace("%usage%",getUsage())));
-                                    break a;
-                                }
-
-                                for (int i = 0; i < instance.getSyntax().getArguments().size(); i++) {
-                                    Argument arg = instance.getSyntax().getArguments().get(i);
-                                    if(arg.getPriority()==ArgumentPriority.REQUIRED){
-                                        if((subArgs.length-1)<i){
-                                            sender.sendMessage(TranslateManager.translateColors(configuration.USAGE.replace("%usage%",getUsage())));
-                                            break a;
-                                        }
-
-                                        if((subArgs[i]==null)){
-                                            sender.sendMessage(TranslateManager.translateColors(configuration.USAGE.replace("%usage%",getUsage())));
-                                            break a;
-                                        }
-                                    }
-                                    if(!(subArgs.length-1<i)&&subArgs[i]!=null){
-                                        ArgumentHandleResult result = arg.handleCorrection(subArgs[i]);
-                                        if(!result.isPassed()){
-                                            sender.sendMessage(TranslateManager.translateColors(configuration.USAGE.replace("%usage%",getUsage()).replace("%reason%",result.getReason())));
-                                            break a;
-                                        }
-                                    }
-                                }
-
-                                instance.onPlayerUse(p,subArgs);
-                            }
-                        }
-
-                    }
-
-            } else if(sender instanceof ConsoleCommandSender){// sender is not a player but the console
-                if(!isPlayerOnly()){ // command is not player only
-
-                    a:{
-                        if(cmdA.playerOnly()){ // SubCommand's @Command is player only
-                            sender.sendMessage(ChatColor.RED+"This command only can be used by a player!");
-                        } else { // SubCommand's @Command is not playeronly
-
-                            Constructor<? extends SubCommand> constructor = cmd.getConstructor(String.class);
-                            constructor.setAccessible(true);
-                            SubCommand instance = constructor.newInstance(cmdA.name());
-
-                            if(subArgs.length==0&&instance.getSyntax().getArguments().stream().anyMatch(argument -> argument.getPriority()==ArgumentPriority.REQUIRED)){
-                                sender.sendMessage(TranslateManager.translateColors(configuration.USAGE.replace("%usage%",getUsage())));
-                                break a;
-                            }
-
-                            for (int i = 0; i < instance.getSyntax().getArguments().size(); i++) {
-                                Argument arg = instance.getSyntax().getArguments().get(i);
-                                if(arg.getPriority()==ArgumentPriority.REQUIRED){
-                                    if((subArgs.length-1)<i){
-                                        sender.sendMessage(TranslateManager.translateColors(configuration.USAGE.replace("%usage%",getUsage())));
-                                        break a;
-                                    }
-
-                                    if((subArgs[i]==null)){
-                                        sender.sendMessage(TranslateManager.translateColors(configuration.USAGE.replace("%usage%",getUsage())));
-                                        break a;
-                                    }
-                                }
-                                if(!(subArgs.length-1<i)&&subArgs[i]!=null){
-                                    ArgumentHandleResult result = arg.handleCorrection(subArgs[i]);
-                                    if(!result.isPassed()){
-                                        sender.sendMessage(TranslateManager.translateColors(configuration.USAGE.replace("%usage%",getUsage()).replace("%reason%",result.getReason())));
-                                        break a;
-                                    }
-                                }
-                            }
-
-                            instance.onConsoleUse((ConsoleCommandSender) sender,subArgs);
-                        }
-                    }
-
-                } else { // command is player only
-
-                    sender.sendMessage(TranslateManager.translateColors(configuration.ONLY_PLAYER));
-
+                if (!cmdA.permission().isEmpty() && !sender.hasPermission(cmdA.permission())) { // SubCommand's @Command has a permission and player don't have the permission
+                    sender.sendMessage(TranslateManager.translateColors(configuration.NO_PERMISSION));
+                    return true;
                 }
+
+                // SubCommand's @Command don't have a permission or player has the permission
+               doExecution(sender,cmdA,cmd,subArgs,configuration);
+
+            } else if (sender instanceof ConsoleCommandSender) {// sender is not a player but the console
+                if (isPlayerOnly()) {
+                    sender.sendMessage(TranslateManager.translateColors(Simpler.getMessageConfiguration().ONLY_PLAYER));
+                    return true;
+                } // command is not player only
+
+                doExecution(sender, cmdA, cmd, subArgs, configuration);
             }
-        } catch (Exception e){
+        } catch (Exception e) {
             e.printStackTrace();
         }
 
         return true;
     }
 
+    private void doExecution(@NotNull CommandSender sender, me.efekos.simpler.commands.Command cmdA, Class<? extends SubCommand> cmd, String[] subArgs, MessageConfiguration configuration) throws NoSuchMethodException, InstantiationException, IllegalAccessException, InvocationTargetException {
+        a:{
+
+            if (cmdA.playerOnly()&&!(sender instanceof Player)) { // SubCommand's @Command is player only
+                sender.sendMessage(TranslateManager.translateColors(configuration.ONLY_PLAYER));
+                return;
+            }
+
+            // SubCommand's @Command is not player-only
+
+            Constructor<? extends SubCommand> constructor = cmd.getConstructor(String.class);
+            constructor.setAccessible(true);
+            SubCommand instance = constructor.newInstance(cmdA.name());
+
+            if (subArgs.length == 0 && instance.getSyntax().getArguments().stream().anyMatch(argument -> argument.getPriority() == ArgumentPriority.REQUIRED)) {
+                sender.sendMessage(TranslateManager.translateColors(configuration.USAGE.replace("%usage%", getUsage())));
+                break a;
+            }
+
+            for (int i = 0; i < instance.getSyntax().getArguments().size(); i++) {
+                Argument arg = instance.getSyntax().getArguments().get(i);
+                if (arg.getPriority() == ArgumentPriority.REQUIRED) {
+                    if ((subArgs.length - 1) < i) {
+                        sender.sendMessage(TranslateManager.translateColors(configuration.USAGE.replace("%usage%", getUsage())));
+                        break a;
+                    }
+
+                    if ((subArgs[i] == null)) {
+                        sender.sendMessage(TranslateManager.translateColors(configuration.USAGE.replace("%usage%", getUsage())));
+                        break a;
+                    }
+                }
+                if (!(subArgs.length - 1 < i) && subArgs[i] != null) {
+                    ArgumentHandleResult result = arg.handleCorrection(subArgs[i]);
+                    if (!result.isPassed()) {
+                        sender.sendMessage(TranslateManager.translateColors(configuration.USAGE.replace("%usage%", getUsage()).replace("%reason%", result.hasReason() ? Objects.requireNonNull(result.getReason()) : "")));
+                        break a;
+                    }
+                }
+            }
+
+            if(sender instanceof Player) instance.onPlayerUse((Player) sender, subArgs);
+            else instance.onConsoleUse((ConsoleCommandSender) sender,subArgs);
+
+        }
+    }
+
+    /**
+     * Returns whether this command has a permission.
+     *
+     * @return Whether this command has a permission.
+     */
+    public boolean hasPermission() {
+        me.efekos.simpler.commands.Command command = this.getClass().getAnnotation(me.efekos.simpler.commands.Command.class);
+        if (command != null) return command.permission() != null;
+        else return false;
+    }
+
     @NotNull
     @Override
     public List<String> tabComplete(@NotNull CommandSender sender, @NotNull String alias, @NotNull String[] args) throws IllegalArgumentException {
-        if(sender instanceof Player){
-            Player p = (Player) sender;
+        if (sender instanceof Player p) {
 
-            if(!p.hasPermission(getPermission()))return new ArrayList<>();
+            if (hasPermission() && !p.hasPermission(Objects.requireNonNull(getPermission()))) return new ArrayList<>();
 
-            if(args.length==1){
+            if (args.length == 1) {
                 ArrayList<String> cmdNames = new ArrayList<>();
-                getSubs().forEach(sub->{
+                getSubs().forEach(sub -> {
                     try {
                         SubCommand cmd = sub.getConstructor(String.class).newInstance(args[0]);
-                        if(p.hasPermission(cmd.getPermission())) cmdNames.add(cmd.getName());
-                    } catch (Exception e){
+                        if (cmd.hasPermission() && p.hasPermission(Objects.requireNonNull(cmd.getPermission())))
+                            cmdNames.add(cmd.getName());
+                    } catch (Exception e) {
                         e.printStackTrace();
                     }
                 });
@@ -307,13 +295,13 @@ public abstract class CoreCommand extends Command {
                 cmdNames.sort(String::compareTo);
                 return cmdNames.stream().filter(s -> s.startsWith(args[0])).collect(Collectors.toList());
 
-            } else if (args.length>1){
+            } else if (args.length > 1) {
                 Class<? extends SubCommand> sub = getSub(args[0]);
-                if(sub==null)return new ArrayList<>();
+                if (sub == null) return new ArrayList<>();
 
                 try {
-                    return sub.getConstructor(String.class).newInstance(args[0]).tabComplete(sender,alias,Arrays.copyOfRange(args,1,args.length));
-                } catch (Exception e){
+                    return sub.getConstructor(String.class).newInstance(args[0]).tabComplete(sender, alias, Arrays.copyOfRange(args, 1, args.length));
+                } catch (Exception e) {
                     e.printStackTrace();
                 }
             }
@@ -325,35 +313,6 @@ public abstract class CoreCommand extends Command {
     @NotNull
     @Override
     public List<String> tabComplete(@NotNull CommandSender sender, @NotNull String alias, @NotNull String[] args, @Nullable Location location) throws IllegalArgumentException {
-        if(sender instanceof Player){
-            Player p = (Player) sender;
-
-            if(!p.hasPermission(getPermission()))return new ArrayList<>();
-
-            if(args.length==1){
-                ArrayList<String> cmdNames = new ArrayList<>();
-                getSubs().forEach(sub->{
-                    try {
-                        SubCommand cmd = sub.getConstructor(String.class).newInstance(args[0]);
-                        if (p.hasPermission(cmd.getPermission())) cmdNames.add(cmd.getName());
-                    } catch (Exception e){
-                        e.printStackTrace();
-                    }
-                });
-                return cmdNames;
-
-            } else if (args.length>1){
-                Class<? extends SubCommand> sub = getSub(args[0]);
-                if(sub==null)return new ArrayList<>();
-
-                try {
-                    return sub.getConstructor(String.class).newInstance(args[0]).tabComplete(sender,alias, Arrays.copyOfRange(args,1,args.length),location);
-                } catch (Exception e){
-                    e.printStackTrace();
-                }
-            }
-        }
-
-        return new ArrayList<>();
+        return tabComplete(sender, alias, args);
     }
 }