+++ src/main/java/me/efekos/simpler/config/MapDataManager.java
@@ -51,7 +51,7 @@ import java.util.*;
  *           <li>Any class that does not contain any type other than the ones above.</li>
  *           </ul>
  */
-public class MapDataManager<K, V> {
+public class MapDataManager<K, V extends Storable> {
 
     /**
      * Main map of all the data stored inside this database.
@@ -132,7 +132,7 @@ public class MapDataManager<K, V> {
     /**
      * Loads all the data from the save before. You don't have to check if file exists, because method does it.
      * @param clazz Accessing a class object of a type parameter is impossible, so you need to give it. Just do
-     * {@code <T>[].class} here, replacing {@code <T>} with name of your type. That type must be same with {@link V}.
+     * {@code Map<K,V>[].class} here, replacing {@code <K>} and {@code <V>} with the name of your type. That type must be same with {@link K} and {@link V}.
      */
     public void load(Class<Map<K,V>> clazz){
         Gson gson = new Gson();
@@ -146,6 +146,8 @@ public class MapDataManager<K, V> {
                 Map<K,V> n = gson.fromJson(reader,clazz);
 
                 data = n;
+
+                reader.close();
             } catch (Exception e){
                 e.printStackTrace();
             }